requirements.txt
flask
flask-cors
python-docx
transformers
numpy
torch
fuzzysearch
stanza
estnltk
gender-guesser
